refactor(backend/runner): move the harness-agnostic turn modules out of execute-cursor (S2 M1)
The eight modules the turn runtime will call, none of which touches
@cursor/sdk, move from src/activities/execute-cursor/ to src/shared/ by
git mv: blueprint-resolver, env-resolver, exact-apply,
extract-structured-output, persist-decision, cost-guard,
attachment-resolver (flat), and workspace-provision, which lands inside
the workspace domain it composes as shared/workspace/session-provision.ts.
Six co-located tests move with their modules; deny-gate-exact-apply stays
in execute-cursor/__tests__ because its subject is the Cursor deny-gate
chain, with exact-apply as one link.

Only import specifiers change. The CloudRepo interface relocates verbatim
from session-lifecycle.ts into blueprint-resolver.ts (its producer), so
the adapter imports the type from shared rather than shared reaching into
the adapter. Every moved file's body, with import lines stripped, is
byte-identical to its origin; that diff was run and is recorded on the
milestone log. The lazy import of extract-structured-output stays lazy.

No golden changed: the hermetic net is 11 files / 17 tests / 17 goldens
before and after. No identifier, comment, log string or behavior changed
in this commit; the names and comments the move makes false are the
next commit's whole content.

Disposition: stigmer-cloud _projects/2026-09/20260911.03.sp.turn-runtime-extraction/tasks/T01_0_plan.md section 3.
