+++ backend/services/runner/src/shared/cost-guard.ts
